fix: rename commands to match other classes
