+++ server/src/application/commands/rates/filter.rates.command.ts
@@ -1,8 +1,8 @@
-import { SeniorityEnum } from "../../domain/enums/seniority.enum";
-import { LanguageEnum } from "../../domain/enums/languaje.enum";
+import { SeniorityEnum } from "../../../domain/enums/seniority.enum";
+import { LanguageEnum } from "../../../domain/enums/languaje.enum";
 
 
-export default class FilterCommandRates {
+export default class FilterRatesCommand {
 
     private technologies: string[];
     private seniority?: SeniorityEnum;
