+++ server/src/application/handlers/rates/update.rate.handler.ts
@@ -1,8 +1,8 @@
 import RateRepository from "../../../infrastructure/repositories/rate.repository";
-import UpdateCommandRates from "../../commands/update.command.rates";
+import UpdateRateCommand from "../../commands/update.command.rates";
 
 class UpdateRateHandler{
-    async execute(command:UpdateCommandRates){
+    async execute(command:UpdateRateCommand){
         const rate = await RateRepository.findOneById(command.getId());
 
         if(!rate){
