+++ server/src/http/actions/create.rate.action.ts
@@ -1,10 +1,10 @@
 import { Request, Response } from "express";
 import CreateRateHandler from "../../application/handlers/rates/create.rate.handler";
-import CreateCommandRates from "../../application/commands/create.command.rates";
+import CreateRateCommand from "../../application/commands/create.command.rates";
 
 class CreateUserAction {
     async run(req: Request, res: Response) {
-        const command: CreateCommandRates = req.body;
+        const command: CreateRateCommand = req.body;
         
         await CreateRateHandler.execute(command);
